+++ python/src/lib.rs
@@ -4,16 +4,14 @@ use pyo3::{prelude::*, wrap_pyfunction, PyClass};
 
 use dbn::{
     compat::{ErrorMsgV1, InstrumentDefMsgV1, SymbolMappingMsgV1, SystemMsgV1},
-    enums::{Compression, Encoding, SType, Schema},
     flags,
     python::EnumIterator,
-    record::{
-        BidAskPair, CbboMsg, ConsolidatedBidAskPair, ErrorMsg, ImbalanceMsg, InstrumentDefMsg,
-        MboMsg, Mbp10Msg, Mbp1Msg, OhlcvMsg, RecordHeader, StatMsg, StatusMsg, SymbolMappingMsg,
-        SystemMsg, TradeMsg,
-    },
-    Metadata, RType, VersionUpgradePolicy, FIXED_PRICE_SCALE, UNDEF_ORDER_SIZE, UNDEF_PRICE,
-    UNDEF_STAT_QUANTITY, UNDEF_TIMESTAMP,
+    Action, BidAskPair, CbboMsg, Compression, ConsolidatedBidAskPair, Encoding, ErrorMsg,
+    ImbalanceMsg, InstrumentClass, InstrumentDefMsg, MatchAlgorithm, MboMsg, Mbp10Msg, Mbp1Msg,
+    Metadata, OhlcvMsg, RType, RecordHeader, SType, Schema, SecurityUpdateAction, Side, StatMsg,
+    StatType, StatUpdateAction, StatusAction, StatusMsg, StatusReason, SymbolMappingMsg, SystemMsg,
+    TradeMsg, TradingEvent, TriState, UserDefinedInstrument, VersionUpgradePolicy, DBN_VERSION,
+    FIXED_PRICE_SCALE, UNDEF_ORDER_SIZE, UNDEF_PRICE, UNDEF_STAT_QUANTITY, UNDEF_TIMESTAMP,
 };
 
 mod dbn_decoder;
@@ -58,13 +56,26 @@ fn databento_dbn(_py: Python<'_>, m: &PyModule) -> PyResult<()> {
     checked_add_class::<StatMsg>(m)?;
     checked_add_class::<CbboMsg>(m)?;
     // PyClass enums
+    checked_add_class::<Action>(m)?;
     checked_add_class::<Compression>(m)?;
     checked_add_class::<Encoding>(m)?;
+    checked_add_class::<InstrumentClass>(m)?;
+    checked_add_class::<MatchAlgorithm>(m)?;
     checked_add_class::<RType>(m)?;
     checked_add_class::<SType>(m)?;
     checked_add_class::<Schema>(m)?;
+    checked_add_class::<SecurityUpdateAction>(m)?;
+    checked_add_class::<Side>(m)?;
+    checked_add_class::<StatType>(m)?;
+    checked_add_class::<StatUpdateAction>(m)?;
+    checked_add_class::<StatusAction>(m)?;
+    checked_add_class::<StatusReason>(m)?;
+    checked_add_class::<TradingEvent>(m)?;
+    checked_add_class::<TriState>(m)?;
+    checked_add_class::<UserDefinedInstrument>(m)?;
     checked_add_class::<VersionUpgradePolicy>(m)?;
     // constants
+    m.add("DBN_VERSION", DBN_VERSION)?;
     m.add("FIXED_PRICE_SCALE", FIXED_PRICE_SCALE)?;
     m.add("UNDEF_PRICE", UNDEF_PRICE)?;
     m.add("UNDEF_ORDER_SIZE", UNDEF_ORDER_SIZE)?;