+++ rust/dbn/src/enums.rs
@@ -15,6 +15,11 @@ use num_enum::{IntoPrimitive, TryFromPrimitive};
 /// A side of the market. The side of the market for resting orders, or the side
 /// of the aggressor for trades.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter, strum::AsRefStr),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[repr(u8)]
 pub enum Side {
     /// A sell order or sell aggressor in a trade.
@@ -33,6 +38,11 @@ impl From<Side> for char {
 
 /// A tick action.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter, strum::AsRefStr),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[repr(u8)]
 pub enum Action {
     /// An existing order was modified.
@@ -57,6 +67,11 @@ impl From<Action> for char {
 
 /// The class of instrument.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[repr(u8)]
 pub enum InstrumentClass {
     /// A bond.
@@ -87,6 +102,11 @@ impl From<InstrumentClass> for char {
 
 /// The type of matching algorithm used for the instrument at the exchange.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[repr(u8)]
 pub enum MatchAlgorithm {
@@ -120,8 +140,13 @@ impl From<MatchAlgorithm> for char {
 
 /// Whether the instrument is user-defined.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, TryFromPrimitive, IntoPrimitive, Default)]
-#[repr(u8)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter, strum::AsRefStr),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[repr(u8)]
 pub enum UserDefinedInstrument {
     /// The instrument is not user-defined.
     #[default]
@@ -139,12 +164,12 @@ impl From<UserDefinedInstrument> for char {
 /// A symbology type. Refer to the [symbology documentation](https://docs.databento.com/api-reference-historical/basics/symbology)
 /// for more information.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, TryFromPrimitive)]
-#[repr(u8)]
 #[cfg_attr(
     feature = "python",
+    derive(strum::EnumIter),
     pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
 )]
-#[cfg_attr(feature = "python", derive(strum::EnumIter))]
+#[repr(u8)]
 pub enum SType {
     /// Symbology using a unique numeric ID.
     InstrumentId = 0,
@@ -222,12 +247,12 @@ pub mod rtype {
 
     /// A type of record, i.e. a struct implementing [`HasRType`](crate::record::HasRType).
     #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, TryFromPrimitive)]
-    #[repr(u8)]
     #[cfg_attr(
         feature = "python",
+        derive(strum::EnumIter),
         pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
     )]
-    #[cfg_attr(feature = "python", derive(strum::EnumIter))]
+    #[repr(u8)]
     pub enum RType {
         /// Denotes a market-by-price record with a book depth of 0 (used for the
         /// [`Trades`](super::Schema::Trades) schema).
@@ -273,11 +298,14 @@ pub mod rtype {
         Mbo = 0xA0,
         /// Denotes a consolidated best bid and offer record.
         Cbbo = 0xB1,
-        /// Denotes a consolidated best bid and offer record subsampled on a one-second interval.
+        /// Denotes a consolidated best bid and offer record subsampled on a one-second
+        /// interval.
         Cbbo1S = 0xC0,
-        /// Denotes a consolidated best bid and offer record subsampled on a one-minute interval.
+        /// Denotes a consolidated best bid and offer record subsampled on a one-minute
+        /// interval.
         Cbbo1M = 0xC1,
-        /// Denotes a consolidated best bid and offer trade record containing the consolidated BBO before the trade.
+        /// Denotes a consolidated best bid and offer trade record containing the
+        /// consolidated BBO before the trade.
         Tcbbo = 0xC2,
         /// Denotes a best bid and offer record subsampled on a one-second interval.
         Bbo1S = 0xC3,
@@ -463,11 +491,14 @@ pub mod rtype {
 
 /// A data record schema.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, TryFromPrimitive)]
-#[repr(u16)]
-#[cfg_attr(feature = "python", pyo3::pyclass(module = "databento_dbn"))]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn")
+)]
 #[cfg_attr(not(feature = "python"), derive(MockPyo3))]
-#[cfg_attr(feature = "python", derive(strum::EnumIter))]
 #[cfg_attr(test, derive(strum::EnumCount))]
+#[repr(u16)]
 pub enum Schema {
     /// Market by order.
     #[pyo3(name = "MBO")]
@@ -609,19 +640,18 @@ impl Display for Schema {
 
 /// A data encoding format.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, TryFromPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[repr(u8)]
-#[cfg_attr(feature = "python", pyo3::pyclass(module = "databento_dbn"))]
-#[cfg_attr(not(feature = "python"), derive(MockPyo3))]
-#[cfg_attr(feature = "python", derive(strum::EnumIter))]
 pub enum Encoding {
     /// Databento Binary Encoding.
-    #[pyo3(name = "DBN")]
     Dbn = 0,
     /// Comma-separated values.
-    #[pyo3(name = "CSV")]
     Csv = 1,
     /// JavaScript object notation.
-    #[pyo3(name = "JSON")]
     Json = 2,
 }
 
@@ -663,16 +693,16 @@ impl Display for Encoding {
 
 /// A compression format or none if uncompressed.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, TryFromPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[repr(u8)]
-#[cfg_attr(feature = "python", pyo3::pyclass(module = "databento_dbn"))]
-#[cfg_attr(not(feature = "python"), derive(MockPyo3))]
-#[cfg_attr(feature = "python", derive(strum::EnumIter))]
 pub enum Compression {
     /// Uncompressed.
-    #[pyo3(name = "NONE")]
     None = 0,
     /// Zstandard compressed.
-    #[pyo3(name = "ZSTD")]
     ZStd = 1,
 }
 
@@ -729,10 +759,15 @@ pub mod flags {
 }
 
 /// The type of [`InstrumentDefMsg`](crate::record::InstrumentDefMsg) update.
-#[repr(u8)]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
-#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 #[allow(clippy::manual_non_exhaustive)] // false positive
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter, strum::AsRefStr),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[repr(u8)]
 pub enum SecurityUpdateAction {
     /// A new instrument definition.
     Add = b'A',
@@ -746,9 +781,14 @@ pub enum SecurityUpdateAction {
 }
 
 /// The type of statistic contained in a [`StatMsg`](crate::record::StatMsg).
-#[repr(u16)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[non_exhaustive]
+#[repr(u16)]
 pub enum StatType {
     /// The price of the first trade of an instrument. `price` will be set.
     OpeningPrice = 1,
@@ -793,8 +833,13 @@ pub enum StatType {
 }
 
 /// The type of [`StatMsg`](crate::record::StatMsg) update.
-#[repr(u8)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
+#[repr(u8)]
 pub enum StatUpdateAction {
     /// A new statistic.
     New = 1,
@@ -803,9 +848,14 @@ pub enum StatUpdateAction {
 }
 
 /// The primary enum for the type of [`StatusMsg`](crate::record::StatusMsg) update.
-#[repr(u16)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive, Default)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[non_exhaustive]
+#[repr(u16)]
 pub enum StatusAction {
     /// No change.
     #[default]
@@ -845,9 +895,14 @@ pub enum StatusAction {
 
 /// The secondary enum for a [`StatusMsg`](crate::record::StatusMsg) update, explains
 /// the cause of a halt or other change in `action`.
-#[repr(u16)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive, Default)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[non_exhaustive]
+#[repr(u16)]
 pub enum StatusReason {
     /// No reason is given.
     #[default]
@@ -879,7 +934,7 @@ pub enum StatusReason {
     NewIssue = 15,
     /// The status changed because an issue is available.
     IssueAvailable = 16,
-    /// The status changed because the issue was reviewed.
+    /// The status changed because the issue(s) were reviewed.
     IssuesReviewed = 17,
     /// The status changed because the filing requirements were satisfied.
     FilingReqsSatisfied = 18,
@@ -924,9 +979,14 @@ pub enum StatusReason {
 }
 
 /// Further information about a status update.
-#[repr(u16)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive, Default)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
 #[non_exhaustive]
+#[repr(u16)]
 pub enum TradingEvent {
     /// No additional information given.
     #[default]
@@ -943,8 +1003,13 @@ pub enum TradingEvent {
 
 /// An enum for representing unknown, true, or false values. Equivalent to
 /// `Option<bool>` but with a human-readable repr.
-#[repr(u8)]
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive, Default)]
+#[cfg_attr(
+    feature = "python",
+    derive(strum::EnumIter),
+    pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
+)]
+#[repr(u8)]
 pub enum TriState {
     /// The value is not applicable or not known.
     #[default]
@@ -979,9 +1044,9 @@ impl From<Option<bool>> for TriState {
 #[derive(Debug, Clone, Copy, Default, PartialEq, Eq, Hash)]
 #[cfg_attr(
     feature = "python",
+    derive(strum::EnumIter),
     pyo3::pyclass(module = "databento_dbn", rename_all = "SCREAMING_SNAKE_CASE")
 )]
-#[cfg_attr(feature = "python", derive(strum::EnumIter))]
 #[non_exhaustive]
 pub enum VersionUpgradePolicy {
     /// Decode data from previous versions as-is.