+++ rust/dbn/src/publishers.rs
@@ -8,6 +8,7 @@ use crate::{Error, Result};
 
 /// A trading execution venue.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[non_exhaustive]
 #[repr(u16)]
 pub enum Venue {
     /// CME Globex
@@ -215,6 +216,7 @@ impl std::str::FromStr for Venue {
 
 /// A source of data.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[non_exhaustive]
 #[repr(u16)]
 pub enum Dataset {
     /// CME MDP 3.0 Market Data
@@ -374,6 +376,7 @@ impl std::str::FromStr for Dataset {
 
 /// A specific Venue from a specific data source.
 #[derive(Clone, Copy, Debug, PartialEq, Eq, Hash, IntoPrimitive, TryFromPrimitive)]
+#[non_exhaustive]
 #[repr(u16)]
 pub enum Publisher {
     /// CME Globex MDP 3.0