+++ CHANGELOG.md
@@ -3,15 +3,29 @@
 ## 0.17.0 - TBD
 
 ### Enhancements
-- Added message type `CbboMsg`, new rtypes and schema types for `Cbbo`, `Cbbo-1s`, `Cbbo-1m`,
-  `Tcbbo`, `Bbo-1s` and `Bbo-1m`
+- Added new record type `CbboMsg`, new rtypes and schema types for `Cbbo`, `Cbbo1s`,
+  `Cbbo1m`, `Tcbbo`, `Bbo1s`, and `Bbo1m`
+- Exported more enums to Python:
+  - `Action`
+  - `InstrumentClass`
+  - `MatchAlgorithm`
+  - `SecurityUpdateAction`
+  - `Side`
+  - `StatType`
+  - `StatUpdateAction`
+  - `StatusAction`
+  - `StatusReason`
+  - `TradingEvent`
+  - `TriState`
+  - `UserDefinedInstrument`
 
 ### Breaking changes
 - Removed `Default` trait implementation for `Mbp1Msg` due to it now having multiple
-  permissible `rtype` values. Users should migrate to call `default_for_schema` instead
+  permissible `rtype` values. Users should use `default_for_schema` instead
 
 ### Bug fixes
 - Fixed an issue where the Python `MappingIntervalDict` was not exported
+- Fixed Python type stubs for `VersionUpgradePolicy` and `SType`
 
 ## 0.16.0 - 2024-03-01
 ### Enhancements
@@ -21,7 +35,7 @@
 - Added `-t` and `--tsv` flags to DBN CLI to encode tab-separated values (TSV)
 - Added `delimiter` method to builders for `DynEncoder` and `CsvEncoder` to customize the
   field delimiter character, allowing DBN to be encoded as tab-separated values (TSV)
-- Document cancellation safety for `AsyncRecordDecoder::decode_ref` (credit: @yongqli)
+- Documented cancellation safety for `AsyncRecordDecoder::decode_ref` (credit: @yongqli)
 - Added new publisher values for consolidated DBEQ.MAX
 - Added C FFI conversion functions from `ErrorMsgV1` to `ErrorMsg` and `SystemMsgV1`
   to `SystemMsg`